services/worker/requirements.txt
celery[redis]
redis
psycopg2-binary
sqlalchemy
python-dotenv
requests
meilisearch

